requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Authlib==1.3.1
requests==2.32.3
email-validator==2.2.0
python-dotenv==1.0.1
